backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pdfplumber==0.10.3
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.2
reportlab==4.0.7
openai==1.3.7
boto3==1.29.7
pydantic==2.5.2
pydantic-settings==2.1.0
bcrypt==4.1.2
PyJWT==2.8.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
opencv-python-headless==4.8.1.78
numpy==1.26.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==7.8.0